+++ src/com/dmdirc/interfaces/ui/ProfilesDialogModelListener.java
@@ -22,7 +22,7 @@
 
 package com.dmdirc.interfaces.ui;
 
-import com.dmdirc.config.profiles.Profile;
+import com.dmdirc.actions.wrappers.Profile;
 
 import java.util.List;
 import java.util.Optional;
